# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XpressPro FX Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets/connect:
    post:
      operationId: connectExternalWallet
      tags: [wallets]
      summary: Connect an external wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectWalletRequest"
      responses:
        "200":
          description: Connected wallet info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectedWallet"
  /wallets/exchange/connect:
    post:
      operationId: connectExchangeWallet
      tags: [wallets]
      summary: Link a MoonPay or Coinbase exchange account as an exchange wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectExchangeWalletRequest"
      responses:
        "200":
          description: Connected exchange wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectedWallet"
  /wallets/exchange/{walletId}:
    delete:
      operationId: disconnectExchangeWallet
      tags: [wallets]
      summary: Remove a previously connected exchange wallet
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /wallets/connected/{walletId}/balance:
    get:
      operationId: getConnectedWalletBalance
      tags: [wallets]
      summary: Fetch live on-chain balance and gas price for a connected wallet
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Live balance information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectedWalletLiveBalance"
  /wallets/connected/{walletId}/send:
    post:
      operationId: sendFromConnectedWallet
      tags: [wallets]
      summary: Sign and broadcast an on-chain transaction from a connected wallet
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendFromConnectedWalletRequest"
      responses:
        "200":
          description: Send transaction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendFromConnectedWalletResult"
...
components:
  schemas:
...
    ConnectWalletRequest:
      type: object
      properties:
        address:
          type: string
          description: Public blockchain address only. Never submit a seed phrase or private key.
        walletType:
          type: string
          description: Wallet provider name (predefined like metamask/trust, or any free-form custom name).
          minLength: 1
          maxLength: 64
      required: [address, walletType]
    ConnectedWallet:
      type: object
      description: |
        Public, user-facing view of a connected external wallet. Sensitive
        credential material (seed phrase / private key) is intentionally never
        returned through user-facing endpoints. Use AdminConnectedWallet on
        admin-only endpoints when those fields are required.
      properties:
        id:
          type: string
        address:
          type: string
        walletType:
          type: string
        balance:
          type: number
        currency:
          type: string
        connectedAt:
          type: string
        provider:
          type: string
          enum: [self_custody, moonpay, coinbase]
          description: Connection class — self_custody for the original Connect-Wallet flow, or an exchange provider for Connect-Exchange-Wallet links.
        label:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
          description: NeXTrade email forwarded to the exchange provider as the user's account identity.
        syncedProfile:
          $ref: "#/components/schemas/SyncedExchangeProfile"
          nullable: true
      required: [id, address, walletType, balance, currency, connectedAt, provider]
    SyncedExchangeProfile:
      type: object
      description: NeXTrade profile fields forwarded to the exchange provider so the user is not re-prompted for sign-up info.
      properties:
        fullName:
          type: string
        email:
          type: string
        country:
          type: string
        phone:
          type: string
          nullable: true
        bankName:
          type: string
          nullable: true
        bankLast4:
          type: string
          nullable: true
        cardLast4:
          type: string
          nullable: true
      required: [fullName, email, country]
    ConnectExchangeWalletRequest:
      type: object
      properties:
        provider:
          type: string
          enum: [moonpay, coinbase]
        address:
          type: string
          description: Public blockchain address only. Never submit a seed phrase or private key.
        label:
          type: string
          nullable: true
      required: [provider, address]
...
    TokenBalance:
      type: object
      properties:
        symbol:
          type: string
        address:
          type: string
        balance:
          type: number
        decimals:
          type: integer
      required: [symbol, address, balance, decimals]
    ConnectedWalletLiveBalance:
      type: object
      properties:
        walletId:
          type: string
        address:
          type: string
        chain:
          type: string
        ethBalance:
          type: number
        tokens:
          type: array
          items:
            $ref: "#/components/schemas/TokenBalance"
        gasPriceGwei:
          type: number
        estimatedSendGasFeeEth:
          type: number
        fetchedAt:
          type: string
        source:
          type: string
          description: Identifies which RPC provider supplied the data (alchemy / infura / public).
        error:
          type: string
          nullable: true
          description: Populated when the live lookup failed (e.g. invalid address, no RPC available).
      required: [walletId, address, chain, ethBalance, tokens, gasPriceGwei, estimatedSendGasFeeEth, fetchedAt, source]
    SendFromConnectedWalletRequest:
      type: object
      properties:
        to:
          type: string
          minLength: 1
        amount:
          type: number
          exclusiveMinimum: 0
        asset:
          type: string
          minLength: 1
          description: Asset symbol to send (ETH or an ERC-20 symbol like USDT/USDC).
      required: [to, amount, asset]
    SendFromConnectedWalletResult:
      type: object
      properties:
        success:
          type: boolean
        hash:
          type: string
          nullable: true
        from:
          type: string
          nullable: true
        to:
          type: string
          nullable: true
        asset:
          type: string
        amount:
          type: number
        message:
          type: string
        blockNumber:
          type: integer
          nullable: true
          description: |
            Block number the transaction was mined in. Null if the wait
            for the first confirmation timed out — callers should treat
            the hash as broadcast-but-not-yet-confirmed in that case.
        confirmations:
          type: integer
          description: |
            Confirmations observed before responding. Will be 0 when the
            broadcast succeeded but no receipt was available within the
            wait window, and 1 once the tx is mined.
        status:
          type: integer
          nullable: true
          description: |
            Receipt status. 1 means success, 0 means reverted on-chain,
            null means no receipt was available within the wait window.
      required: [success, asset, amount, message, confirmations]
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...